+++ src/InstaSharp/Endpoints/Relationships.cs
@@ -83,7 +83,7 @@ namespace InstaSharp.Endpoints
         /// </summary>
         /// <param name="userId">The list of users that this user id is following.</param>
         /// <returns>UsersResponse</returns>
-        public Task<UsersResponse> Follows(int userId)
+        public Task<UsersResponse> Follows(long userId)
         {
             return Follows(userId, null);
         }
@@ -96,7 +96,7 @@ namespace InstaSharp.Endpoints
         /// <param name="userId">The list of users that this user id is following.</param>
         /// <param name="cursor">The next cursor id</param>
         /// <returns>UsersResponse</returns>
-        public Task<UsersResponse> Follows(int userId, string cursor)
+        public Task<UsersResponse> Follows(long userId, string cursor)
         {
             var request = Request("{id}/follows");
             request.AddUrlSegment("id", userId.ToString());
@@ -150,7 +150,7 @@ namespace InstaSharp.Endpoints
         /// </summary>
         /// <param name="userId">The id of the user to get the followers of.</param>
         /// <returns>Users response</returns>
-        public Task<UsersResponse> FollowedBy(int userId)
+        public Task<UsersResponse> FollowedBy(long userId)
         {
             return FollowedBy(userId, null);
         }
@@ -163,7 +163,7 @@ namespace InstaSharp.Endpoints
         /// <param name="userId">The id of the user to get the followers of.</param>
         /// <param name="cursor">The next cursor id</param>
         /// <returns>Users response</returns>
-        public Task<UsersResponse> FollowedBy(int userId, string cursor)
+        public Task<UsersResponse> FollowedBy(long userId, string cursor)
         {
             var request = Request("{id}/followed-by");
             request.AddUrlSegment("id", userId.ToString());
